requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
faiss-cpu==1.13.2
filelock==3.25.2
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
nltk==3.9.3
numpy==2.4.3
openpyxl==3.1.5
packaging==26.0
pillow==12.1.1
Pygments==2.19.2
PyMuPDF==1.27.2
python-docx==1.2.0
python-pptx==1.0.2
PyYAML==6.0.3
rank-bm25==0.2.2
regex==2026.2.28
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
setuptools==82.0.1
shellingham==1.5.4
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0+cpu
torchaudio==2.10.0+cpu
torchvision==0.25.0+cpu
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing_extensions==4.15.0
watchdog==6.0.0
xlsxwriter==3.2.9
